+++ src/drv/led/init.cpp
@@ -31,38 +31,38 @@ namespace drv::led {
 /// timer with 2 channels for the bug and WiFi LEDs. The resolution of the PWM
 /// is 8 bits.
 esp_err_t init() {
-  static constexpr ledc_timer_config_t timer{.speed_mode = LEDC_LOW_SPEED_MODE,
-                                             .duty_resolution =
-                                               LEDC_TIMER_8_BIT,
-                                             .timer_num = LEDC_TIMER_0,
-                                             .freq_hz = 4000,
-                                             .clk_cfg = LEDC_AUTO_CLK};
-  ESP_ERROR_CHECK(ledc_timer_config(&timer));
+  static constexpr ledc_timer_config_t timer_cfg{
+    .speed_mode = LEDC_LOW_SPEED_MODE,
+    .duty_resolution = LEDC_TIMER_8_BIT,
+    .timer_num = LEDC_TIMER_0,
+    .freq_hz = 4000,
+    .clk_cfg = LEDC_AUTO_CLK};
+  ESP_ERROR_CHECK(ledc_timer_config(&timer_cfg));
 
   {
-    static constexpr ledc_channel_config_t cfg{.gpio_num = bug_gpio_num,
-                                               .speed_mode =
-                                                 LEDC_LOW_SPEED_MODE,
-                                               .channel = bug_channel,
-                                               .intr_type = LEDC_INTR_DISABLE,
-                                               .timer_sel = LEDC_TIMER_0,
-                                               .duty = 0u,
-                                               .hpoint = 0};
-    ESP_ERROR_CHECK(ledc_channel_config(&cfg));
+    static constexpr ledc_channel_config_t channel_cfg{
+      .gpio_num = bug_gpio_num,
+      .speed_mode = LEDC_LOW_SPEED_MODE,
+      .channel = bug_channel,
+      .intr_type = LEDC_INTR_DISABLE,
+      .timer_sel = LEDC_TIMER_0,
+      .duty = 0u,
+      .hpoint = 0};
+    ESP_ERROR_CHECK(ledc_channel_config(&channel_cfg));
     ESP_ERROR_CHECK(ledc_set_duty(LEDC_LOW_SPEED_MODE, bug_channel, 0u));
     ESP_ERROR_CHECK(ledc_update_duty(LEDC_LOW_SPEED_MODE, bug_channel));
   }
 
   {
-    static constexpr ledc_channel_config_t cfg{.gpio_num = wifi_gpio_num,
-                                               .speed_mode =
-                                                 LEDC_LOW_SPEED_MODE,
-                                               .channel = wifi_channel,
-                                               .intr_type = LEDC_INTR_DISABLE,
-                                               .timer_sel = LEDC_TIMER_0,
-                                               .duty = 0u,
-                                               .hpoint = 0};
-    ESP_ERROR_CHECK(ledc_channel_config(&cfg));
+    static constexpr ledc_channel_config_t channel_cfg{
+      .gpio_num = wifi_gpio_num,
+      .speed_mode = LEDC_LOW_SPEED_MODE,
+      .channel = wifi_channel,
+      .intr_type = LEDC_INTR_DISABLE,
+      .timer_sel = LEDC_TIMER_0,
+      .duty = 0u,
+      .hpoint = 0};
+    ESP_ERROR_CHECK(ledc_channel_config(&channel_cfg));
     ESP_ERROR_CHECK(ledc_set_duty(LEDC_LOW_SPEED_MODE, wifi_channel, 0u));
     return ledc_update_duty(LEDC_LOW_SPEED_MODE, wifi_channel);
   }