+++ src/drv/out/track/zimo/mdu/resume.cpp
@@ -25,9 +25,9 @@
 namespace drv::out::track::zimo::mdu {
 
 /// \todo document
-esp_err_t init_encoder(mdu_encoder_config_t const& encoder_config) {
+esp_err_t init_encoder(mdu_encoder_config_t const& encoder_cfg) {
   assert(!encoder);
-  return rmt_new_mdu_encoder(&encoder_config, &encoder);
+  return rmt_new_mdu_encoder(&encoder_cfg, &encoder);
 }
 
 namespace {
@@ -49,9 +49,9 @@ esp_err_t init_gpio(gpio_isr_t gpio_isr_handler) {
 } // namespace
 
 /// \todo document
-esp_err_t resume(mdu_encoder_config_t const& encoder_config,
+esp_err_t resume(mdu_encoder_config_t const& encoder_cfg,
                  gpio_isr_t gpio_isr_handler) {
-  ESP_ERROR_CHECK(init_encoder(encoder_config));
+  ESP_ERROR_CHECK(init_encoder(encoder_cfg));
   ESP_ERROR_CHECK(init_alarm());
   return init_gpio(gpio_isr_handler);
 }