+++ src/intf/dns/init.cpp
@@ -37,16 +37,15 @@ namespace intf::dns {
 /// The DNS server will **not** redirect HTTPS requests.
 esp_err_t init() {
   // Only start DNS in AP mode
-  wifi_mode_t mode;
-  ESP_ERROR_CHECK(esp_wifi_get_mode(&mode));
+  wifi_mode_t mode{WIFI_MODE_NULL};
+  esp_wifi_get_mode(&mode);
 
   // Start the DNS server that will redirect all queries to the softAP IP
   if (mode == WIFI_MODE_AP) {
-    dns_server_config_t config =
+    dns_server_config_t cfg =
       DNS_SERVER_CONFIG_SINGLE("*",            // all A queries
                                "WIFI_AP_DEF"); // softAP netif ID
-
-    start_dns_server(&config);
+    start_dns_server(&cfg);
   }
 
   return ESP_OK;